# Optimize RDT component picker setup perf