@markerikson 2023-10-24
Note that our RDT E2E tests should indeed try both the "fast path" and "slow path" - an FF recording for the slow path, and a Chromium recording past 2023-09-16 for the fast path.